Move DictionaryReader to parquet tests
