+++ lib/trino-parquet/src/test/java/io/trino/parquet/reader/decoders/DictionaryReader.java
@@ -11,8 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.parquet.dictionary;
+package io.trino.parquet.reader.decoders;
 
+import io.trino.parquet.dictionary.Dictionary;
 import org.apache.parquet.bytes.ByteBufferInputStream;
 import org.apache.parquet.bytes.BytesUtils;
 import org.apache.parquet.column.values.ValuesReader;
